+++ src/main/java/com/procedures/demo/user/UserRepository.java
@@ -2,4 +2,4 @@ package com.procedures.demo.user;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface UserRepository extends CrudRepository<User, Integer>{ }
+public interface UserRepository extends CrudRepository<User, Long>{ }
